move time->clock
