+++ lib/time/clock/am335x/BUILD
@@ -4,14 +4,14 @@ package(default_visibility = ["//visibility:public"])
 cc_library(
     name = "am335x",
     hdrs = [
-        "time-am335x.h",
+        "clock-am335x.h",
     ],
     srcs = [
-        "time-am335x.c",
+        "clock-am335x.c",
     ],
     deps = [
         "//lib/thread",
-        "//lib/time:defs",
+        "//lib/time/clock:defs",
         "//tools/toolchain/am335x",
     ],
 )
