+++ lib/time/clock/clock.h
@@ -10,12 +10,12 @@
 extern "C" {
 #endif
 
-#include "lib/time/time-defs.h"
+#include "lib/time/clock/clock-defs.h"
 
 #if defined(SUPRUGLUE_AM335X)
-#include "lib/time/am335x/time-am335x.h"
+#include "lib/time/clock/am335x/clock-am335x.h"
 #elif defined(SUPRUGLUE_TEST32)
-#include "lib/time/test32/time-test32.h"
+#include "lib/time/clock/test32/clock-test32.h"
 #endif
 
 void Sleep(uint32_t nanos);
