+++ lib/time/clock/test32/clock-test32.cc
@@ -1,4 +1,4 @@
 // Copyright Joshua MacDonald
 // SPDX-License-Identifier: MIT
 
-#include "lib/time/time-defs.h"
+#include "lib/time/clock/clock-defs.h"
