+++ lib/time/clock/test32/clock-test32.h
@@ -8,7 +8,7 @@
 extern "C" {
 #endif
 
-#include "lib/time/time-defs.h"
+#include "lib/time/clock/clock-defs.h"
 
 #ifdef __cplusplus
 }
